// anime-web/types/anime.ts
export interface ApiResponse<T> {
  statusCode: number;
  statusMessage: string;
  message: string;
  ok: boolean;
  data: T;
  pagination: Pagination | null;
}
 
export interface Pagination {
  currentPage: number;
  hasPreviousPage: boolean;
  prevPage: number | null;
  hasNextPage: boolean;
  nextPage: number | null;
  totalPages: number;
}
// ...
export interface AnimeDetail {
    title: string;
    poster: string;
    score: {
        value: string;
        users: string;
    }
    japanese: string;
    synonyms: string[];
    english: string;
    status: string;
    type: string;
    source: string;
    duration: string;
    episodes: string;
    season: string;
    studios: string[];
    producers: string[];
    aired: string;
    trailer: string;
    synopsis: {
        paragraphs: string;
        connections: string[];
    }
    genreList: {
        title: string;
        genreId: string;
        href: string;
        samehadakuUrl: string;
    }[];
    batchList: {
        title: string;
        batchId: string;
        href: string;
        samehadakuUrl: string;
    } [];
    episodeList: {
        title: string;
        episodeId: string;
        href: string;
        samehadakuUrl: string;
    }[];
}